Adding Searchbar functionality in invoices page.
Adding search bar functionality in the invoices page using client hooks:
useSearchParams, usePathname, useRouter. When the user types a parameter
in the search bar, the table will automatically update to reflect the
search parameter. To prevent a new database query on every keystroke,
useDebouncedCallback was used to only query the database after the user
has stopped typing for 300 ms.